[3. ネットワークの準備]学習の実施
- 学習実施部分の作成
- 損失関数を変化するloss_plotを作成

+++ code/train_value_dataset.py
@@ -11,7 +11,7 @@ class Train_value_dataset(Dataset):
         return len(self.X)
 
     def __getitem__(self, idx):
-        return self.X[idx], self.y[idx]
+        return torch.FloatTensor(self.X[idx]), self.y[idx]
 
 
 def main():
